refactor(services): provider account helper를 spec 파일로 분리 (#50 step C)
codex-provider / claude-provider에 인라인으로 있던 account filter/map 함수를
provider별 spec 파일로 추출.

- services/codex-account-spec.js: filterRealCodexAccounts, codexMapAccountToProfile
- services/claude-account-spec.js: filterClaudeRealAccounts, claudeMapAccountToProfile, matchesFilter

provider 파일은 spec을 import해서 resolveProviderProfiles에 주입만 한다.
기존 re-export(filterRealCodexAccounts 등)는 유지.

489 pass.

refs #50
